% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_capital.R
\name{calcCapitalAllocation}
\alias{calcCapitalAllocation}
\title{Develop Capital Allocation Strategy for Cyber Risk Management}
\usage{
calcCapitalAllocation(
  invest.table,
  cyra.weights,
  normalizer = "cte",
  cyra.params,
  scale.fun = scaleLogNormal,
  penalty.thres = 0.9,
  ...
)
}
\arguments{
\item{invest.table}{a data.frame of class \code{cyraInvestTable} generated by function \code{\link[=createInvestTable]{createInvestTable()}}.}

\item{cyra.weights}{a list of class \code{cyraWeights} generated by function \code{\link[=createWeightsTables]{createWeightsTables()}}.}

\item{normalizer}{either a character value that specifies how the expected deviance should be normalized, or a custom function that returns a vector of normalizing factors.}

\item{cyra.params}{a list of class \code{cyraParams} generated by function \code{\link[=prepareCyraParameters]{prepareCyraParameters()}}.}

\item{scale.fun}{a function that scales a random variable. \code{\link[=scaleLogNormal]{scaleLogNormal()}} is used by default, and should be changed accordingly depending on the distribution fitted to losses.}

\item{penalty.thres}{a numeric value between 0 and 1, which is the probability used in the TVaR penalty function. 0.9 is used by default.}

\item{...}{additional arguments for \code{normalizer} if it is a custom function.}
}
\value{
a data.frame that shows all cybersecurity investment options and all corresponding optimal reserving schemes.
\itemize{
\item \code{investj} for \code{j}\eqn{=1,2,\dots,m}: amount spent on control/vulnerability \code{j}.
\item \code{thetaj} for \code{j}\eqn{=1,2,\dots,m}: theta value after the cybersecurity investment is made on control/vulnerability \code{j}.
\item \code{Kik} for \code{i}\eqn{=1,2,\dots,l} and \code{k}\eqn{=1,2,\dots,n}: reserve on threat-asset pair \code{ik}.
\item \code{obj_tot}: objective function value of the corresponding investment and reserving strategy.
\item \code{obj_c}: component in the objective function that is related to the investment part.
\item \code{obj_r}: component in the objective function that is related to the reserving part.
}
}
\description{
This function outputs capital allocation strategies with discrete investment options according to the holistic principle. This function should only be used when all components of the allocation model are ready.
}
