% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conditional_expection.R
\name{calcConditionalExpection}
\alias{calcConditionalExpection}
\title{Calculate Conditional Expectation}
\usage{
calcConditionalExpection(x, p, q)
}
\arguments{
\item{x}{a numeric value representing the Value-at-Risk of random variable \eqn{X}}

\item{p}{a numeric vector of probabilities of the distribution of \eqn{X}}

\item{q}{a numeric vector of quantiles corresponding to \code{p}}
}
\description{
This function calculates the conditional expectation of \eqn{X}, when the Value-at-Risk of \eqn{X} is given
}
