% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_normalizer.R
\name{calcNormalizer.l1norm}
\alias{calcNormalizer.l1norm}
\alias{calcNormalizer.local}
\alias{calcNormalizer.cte}
\title{Built-in Weight Normalizers}
\usage{
calcNormalizer.l1norm(term.pt = "TVaR", h.func, p, q, ...)

calcNormalizer.local(h.func, p, q, ...)

calcNormalizer.cte(p, q, ...)
}
\arguments{
\item{term.pt}{either a numerical value or the character value \code{"TVaR"} (default). This is the terminal point such that the l1-norm of the quadratic function and that of the linear function are equal with in \link[=0, `term.pt`]{0, \code{term.pt}}. If \code{term.pt = "TVaR"}, then the value of the terminal point is the TVaR of the loss.  This argument is needed only for function \code{calcNormalizer.l1norm}.}

\item{h.func}{a weighted penalty function. See \code{\link[=calcPenaltyValue.CTE]{calcPenaltyValue.CTE()}}.}

\item{p}{a numeric vector representing the probabilities of the loss random variable in the expected deviance.}

\item{q}{a numeric vector that has the same length as \code{p} and represent the quantiles corresponding to \code{p}.}

\item{...}{additional needed arguments in \code{h.func}.}
}
\value{
a numeric value of the calculated normalizer.
}
\description{
These functions are possible choices of the normalizer of each expected deviance term.
}
\section{Functions}{
\itemize{
\item \code{calcNormalizer.l1norm}: This function calculates the normalizer that equates the l1-norm of the expected deviance to that of the linear reserve term, up to a terminal point.

\item \code{calcNormalizer.local}: This function calculates the normalizer that equates the expected deviance to the linear reserve term at the point where the expected deviance is minimized.

\item \code{calcNormalizer.cte}: This function calculates the normalizer that equates the average marginal decrease in expected deviance to the (marginal) increase in linear reserve term within the interval from no reserve to the amount that minimizes the expected deviance. In that case, the normalizer takes the value of the CTE of the loss random varible.
}}

