% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_penalty.R
\name{calcPenaltyValue.CTE}
\alias{calcPenaltyValue.CTE}
\title{Weighted Penalty Function - CTE}
\usage{
calcPenaltyValue.CTE(x, p, q)
}
\arguments{
\item{x}{a numeric value representing the parameter \eqn{x} in the weighted penalty function.}

\item{p}{a numeric vector that represents the probabilities of the loss distribution.}

\item{q}{a numeric vector that has the same length as \code{p} and represents the quantiles corresponding to \code{p}.}
}
\value{
a numeric value representing the penalty
}
\description{
\loadmathjax This is a the weighted penalty function defined as follows, \mjseqn{\frac{1_{\{X > x\}}}{\mathbb{P}(X > x)}}.
}
