% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_capital.R
\name{calcPenaltyValue}
\alias{calcPenaltyValue}
\title{Develop Capital Allocation Strategy for Cyber Risk Management}
\usage{
calcPenaltyValue(x, p, q)
}
\description{
This function outputs capital allocation strategies with discrete investment options according to the holistic principle
}
