% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_VaR.R
\name{calcVaR}
\alias{calcVaR}
\title{Calculate Value-at-Risk of a Discretized Distribution}
\usage{
calcVaR(alpha, p, q)
}
\arguments{
\item{alpha}{a numeric value representing the probability corresponding to the VaR to be found}

\item{p}{a numeric vector of probabilities of the distribution of \eqn{X}}

\item{q}{a numeric vector of quantiles corresponding to \code{p}}
}
\description{
This function calculates the Value-at-Risk a discretized version of \eqn{X}
}
