% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{cyraSampleData}
\alias{cyraSampleData}
\title{cyra Sample Data for Demonstration and Testing}
\format{
An object of class \code{data.frame} with 1778 rows and 6 columns.
}
\usage{
data(cyraSampleData)
}
\description{
This is a dataset generated by function \code{\link[=generateSyntheticData]{generateSyntheticData()}}
}
\keyword{datasets}
