% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_distribution.R
\name{generateDiscretizedProbs}
\alias{generateDiscretizedProbs}
\title{Discretize Continuous Loss Distributions}
\usage{
generateDiscretizedProbs(q, pdist, ...)
}
\arguments{
\item{q}{a numeric vector of quantiles, for which the discretized probabilites are calculated}

\item{pdist}{a character value, indicating the name of the continuous distribution, e.g., "lnorm", "weibull", etc.}

\item{...}{a named numeric vector of parameters of the distribution.}
}
\value{
a numeric vector of probabilities that corresponds to the input quantiles \code{q}.
}
\description{
This function discretizes a continuous distribution by allocating the probability within each interval to the beginning of that interval
}
