% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_risk.R
\name{generateLossDists}
\alias{generateLossDists}
\title{Aggregate Loss Distribution Functions}
\usage{
generateLossDists(th, cyra.params, scale.fun = scaleLogNormal)
}
\arguments{
\item{th}{a numeric vector, of which the length equals the number of vulnerabilities. Elements in the vector range from 0 to 1. \code{th} (Vector \eqn{\theta}) represents the scaling of loss random variables when cybersecurity controls are present. If losses are unmitigated, then \code{th} should be a vector of one's.}

\item{cyra.params}{a list of S3 class cyraParams, generated by function \code{\link[=prepareCyraParameters]{prepareCyraParameters()}}.}

\item{scale.fun}{a function that returns the new parameters of a probability distribution of a random variable after it is scaled. For more details, see built-in functions \code{\link[=scaleLogNormal]{scaleLogNormal()}} (default) and \code{\link[=scaleWeibull]{scaleWeibull()}}. A compatible customer function can also be provided.}
}
\value{
This function returns a list of discretized probability distributions that are represented by quantile-probability tables. Items in the list are as follows,
\itemize{
\item X_dist: the distribution of \eqn{X_{ijk}}, the loss random variable on each attack path defined by a threat-vulnerability-asset tuple.
\item Zik_dist: the distribution of \eqn{Z_{ik}}, the loss random variable associated with each threat-asset pair.
\item Zi_dist: the distribution of \eqn{Z_i}, the loss random variable associated with each threat.
\item L_dist: the distribution of \eqn{L}, the random variable of the total loss in one incident.
\item S_dist: the distribution of \eqn{S}, the random variable of the aggregate loss over a period.
\item Sik_dist: the distribution of \eqn{S_{ik}}, the random variable associated with the aggregate loss on each threat-asset pair over a period.
}
}
\description{
The function generates discretized probability distributions of various losses arising from a cyber system.
}
