% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{generateSyntheticData}
\alias{generateSyntheticData}
\title{Generate Synthetic Industry Cyber Incident Data}
\usage{
generateSyntheticData(attack.paths, lambda, years, prob.com, seed = 199)
}
\arguments{
\item{attack.paths}{a list of attack path information, see \code{Details}.}

\item{lambda}{a number, which is parameter of the Poisson distribution used for generating yearly number of incidents in this industry.}

\item{years}{a numeric vector that represents the time span of the sample.}

\item{prob.comp}{a number, which represents probability of the incident occurring to the company of interest.}

\item{seed}{a number, seed of random number generator.}
}
\value{
This function returns a cyraIncidents class data.frame of synthetic cyber incidents. Its structure is compatible with the risks assessment and capital allocation framework in this package.
}
\description{
This function generates synthetic industry cyber incident data by simulation. The generated dataset is compatible with the risk assessment and capital allocation framework proposed by Chong, et al. and is for demonstration and testing purposes only.
}
\details{
To be added
}
\examples{
# create attack paths information
path_info <- list(
list("T" = 1, "V" = 1,"A" = 1, "Xzero" = 0.90, "Xdist" = "lnorm", "Xparams" = c("meanlog" = 10, "sdlog" = 5), 'prob' = 0.3),
list("T" = 2, "V" = 2,"A" = 2, "Xzero" = 0.92, "Xdist" = "lnorm", "Xparams" = c("meanlog" = 2, "sdlog" = 10), 'prob' = 0.5),
list("T" = 2, "V" = 3,"A" = 2, "Xzero" = 0.85, "Xdist" = "lnorm", "Xparams" = c("meanlog" = 5, "sdlog" = 1), 'prob' = 0.2)
)
lambda <- 100
years <- 2001:2008
prob_comp <- 0.1
sample_data <- generateSyntheticData(path_info = path_info, lambda = lambda, years = years, prob_comp = prob_comp)
}
