% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{prepareCyraParameters}
\alias{prepareCyraParameters}
\title{Estimate Parameters for Risk Assessment and Capital Allocation}
\usage{
prepareCyraParameters(
  ind.data,
  loss.dist.name = "lnorm",
  grid.ub.prob = 0.95,
  grid.num = 1000L
)
}
\arguments{
\item{ind.data}{A cyraIncidents class data.frame.}

\item{loss.dist.name}{A character value, specifies the name of the distribution that will be fitted to the loss data. Common choices are "lnorm" (by defauly, lognormal distribution) and "weibull" (Weibull distribution).}

\item{grid.ub.prob}{A numeric value, specifies the probability corresponding to the largest quantile that will be used in the process of discretizing loss distributions.}

\item{grid.num}{An integer value, specifies the number of grids when discretizing loss distributions.}
}
\value{
This function returns a list of parameters with their values, which will be used for risk assessment and capital allocation.
}
\description{
Based on the dataset of industry cyber incidents, this functions estimates all parameters that are needed for cyber risk assessment and capital allocation.
}
\examples{
risk_params <- generateRiskParameters(ind.data = sample_data)
}
