% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_distribution.R
\name{scaleLogNormal}
\alias{scaleLogNormal}
\alias{scaleWeibull}
\title{Scale a Random Variable}
\usage{
scaleLogNormal(th, meanlog, sdlog)

scaleWeibull(th, shape, scale)
}
\description{
Here are two built-in functions for scaling lognormal-distributed and Weibull-distributed random variables.
}
\section{Functions}{
\itemize{
\item \code{scaleLogNormal}: scales a lognormal-distributed random variable, i.e., \eqn{Y = \theta X}, where X follows a lognormal distribution with parameters \code{meanlog} and \code{sdlog}. The function outputs the parameters of Y.

\item \code{scaleWeibull}: scales a Weibull-distributed random variable, i.e., \eqn{Y = \theta X}, where X follows a Weibull distribution with parameters \code{scale} and \code{shape}. The function outputs the parameters of Y.
}}

