% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_invest_options.R
\name{createInvestTable}
\alias{createInvestTable}
\alias{cyraInvestOptions}
\alias{cyraInvestOption}
\alias{extractFromInvestTable}
\title{Create Cyber Control Investment Table with Discrete Investment Options}
\usage{
createInvestTable(inv.opt)

cyraInvestOptions(...)

cyraInvestOption(idx, invest, theta)

extractFromInvestTable(invest.table, field)
}
\arguments{
\item{inv.opt}{a list of class \code{cyraInvestOptions} generated by function \code{\link[=cyraInvestOptions]{cyraInvestOptions()}}.}

\item{...}{objects of class \code{cyraInvestOption}}

\item{idx}{an integer indicating the index of the vulnerability that this investment option is for.}

\item{invest}{a numeric value representing the amount spent on this vulerability}

\item{theta}{a numeric value between 0 and 1, indicating the remaining vulnerability after the investment is made.}

\item{invest.table}{a \code{cyraInvestTable} object.}

\item{field}{a character value, of which the value is either \code{invest} or \code{theta}}
}
\description{
Assume that there are discrete investment options on individual cybersecurity vulnerabilities, these functions help populates all possible combinations of investment decisions on individual vulnerabilities.
}
\section{Functions}{
\itemize{
\item \code{cyraInvestOptions}: this function creates a list of class \code{cyraInvestOptions}, which is used in function \code{\link[=createInvestTable]{createInvestTable()}}.

\item \code{cyraInvestOption}: this function creates an individual \code{cyraInvestOption} item.

\item \code{extractFromInvestTable}: this is a helper function that extracts \code{invest} or \code{theta} information from a \code{cyraInvestTable} object (see \code{\link[=createInvestTable]{createInvestTable()}}).
}}

