% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_weights.R
\name{createWeightsTables}
\alias{createWeightsTables}
\alias{createNuOmegaikTable}
\alias{createEtaTable}
\alias{createNuOmegaTable}
\title{Prepare Weights for Capital Allocation}
\usage{
createWeightsTables(cyra.params, nu.ik, om.ik, eta.j, nu, om, eta)

createNuOmegaikTable(l, n, nu.ik, om.ik)

createEtaTable(m, eta.j, eta)

createNuOmegaTable(nu, om)
}
\arguments{
\item{nu.ik, om.ik}{numeric vectors specifying standalone reserve weights, nu_ik and omega_ik.}

\item{eta.j}{a numeric vector specifying standalone cybersecurity investment weights.}

\item{nu, om}{numeric values specifying aggregate reserve weights, nu and omega.}

\item{eta}{a numeric value specifying aggregate cybersecurity investment weights.}

\item{l, m, n}{integer values. l: number of threats; m: number of vulnerabilities; n: number of assets.}
}
\value{
a list of weights tables
\itemize{
\item \code{nu_om_ik}: weights on reserves (\eqn{\nu_{ik}}) and reserve-loss mismatches (\eqn{\omega_{ik}}) on individual threat-asset pairs.
\item \code{nu_om}: weights on aggregate reserve (\eqn{\nu}) and aggregate reserve-loss mismatch (\eqn{\omega})
\item \code{etas}: weights on cybersecurity investments, including standalone investments on individual controls and the aggregate investment.
}
}
\description{
This function can be used to format the user-provided weights into data.frames that are compatible with the risk assessment and capital allocation framework.
}
\section{Functions}{
\itemize{
\item \code{createNuOmegaikTable}: create a table for standalone reserve weights, nu_ik and omega_ik

\item \code{createEtaTable}: create a compatible table for investment weights, eta

\item \code{createNuOmegaTable}: create a compatible table for aggregate reserve weights, nu and omega
}}

