% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_risk.R
\name{generateLossDists}
\alias{generateLossDists}
\title{Aggregate Loss Distribution Functions}
\usage{
generateLossDists(th, cyra.params, scale.fun = scaleLogNormal)
}
\description{
The function generate a discretized version of the aggregate loss (S) distribution and distributions of losses on individual threat-asset pairs (S_{ik}).
}
